pyproject.toml
[tool.poetry]
name = "subscree"
version = "0.1.0"
description = ""
authors = ["Dennis Wainaina <dennis@byteslab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
sqlalchemy = "^2.0.3"
celery = "^5.2.7"
jinja2 = "^3.1.2"
alembic = "^1.9.3"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
psycopg2 = "^2.9.5"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^4.5.1"
shortuuid = "^1.0.11"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
httpx = "^0.23.3"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
